Use compare_state_dicts everywhere
